+++ src/gui/XournalView.cpp
@@ -39,13 +39,8 @@ XournalView::XournalView(GtkWidget* parent, Control* control, ScrollHandling* sc
 
     registerListener(control);
 
-    InputContext* inputContext = nullptr;
-    if (this->control->getSettings()->getExperimentalInputSystemEnabled()) {
-        inputContext = new InputContext(this, scrollHandling);
-        this->widget = gtk_xournal_new(this, inputContext);
-    } else {
-        this->widget = gtk_xournal_new_deprecated(this, scrollHandling);
-    }
+    InputContext* inputContext = new InputContext(this, scrollHandling);
+    this->widget = gtk_xournal_new(this, inputContext);
     // we need to refer widget here, because we unref it somewhere twice!?
     g_object_ref(this->widget);